# packages/api-client/openapi.json
{
  "components" : {
    "schemas" : {
      "ApiProblem" : {
        "description" : "RFC 7807 problem detail",
        "properties" : {
          "detail" : {
            "type" : [ "string", "null" ]
          },
          "errors" : {
            "description" : "Field errors: type urn:tekpas:problem:validation, or conflict on a unique field (code Unique)",
            "items" : {
              "$ref" : "#/components/schemas/FieldViolation"
            },
            "type" : [ "array", "null" ]
          },
          "instance" : {
            "type" : [ "string", "null" ]
          },
          "reason" : {
            "description" : "Conflicts that are not about one field, e.g. PRODUCT_HAS_BATCHES, GTIN_LOCKED",
            "example" : "PRODUCT_HAS_BATCHES",
            "type" : [ "string", "null" ]
          },
          "requestId" : {
            "description" : "Id of the request, also in the X-Request-Id header and the server log",
            "example" : "ab12-cd34",
            "type" : [ "string", "null" ]
          },
          "status" : {
            "example" : 400,
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "example" : "Validation failed",
            "type" : "string"
          },
          "type" : {
            "example" : "urn:tekpas:problem:validation",
            "type" : "string"
          }
        },
        "required" : [ "status", "title", "type" ]
      },
      "BatchCreateRequest" : {
        "properties" : {
          "batchNo" : {
            "example" : "KP-2026-1003-A",
            "pattern" : "^[A-Z0-9-]{0,20}$",
            "type" : [ "string", "null" ]
          },
          "producedFrom" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "producedTo" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "productId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "productionOrderNo" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          },
          "quantity" : {
            "exclusiveMinimum" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "productId", "quantity" ],
        "type" : "object"
      },
      "BatchProduct" : {
        "properties" : {
          "gtin" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "gtin", "id", "name" ],
        "type" : "object"
      },
      "BatchResponse" : {
        "properties" : {
          "batchNo" : {
            "type" : "string"
          },
          "chain" : {
            "$ref" : "#/components/schemas/ChainSummary"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "producedFrom" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "producedTo" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "product" : {
            "$ref" : "#/components/schemas/BatchProduct"
          },
          "productionOrderNo" : {
            "type" : [ "string", "null" ]
          },
          "quantity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "$ref" : "#/components/schemas/BatchStatus"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "batchNo", "chain", "createdAt", "id", "product", "quantity", "status", "updatedAt" ],
        "type" : "object"
      },
      "BatchStatus" : {
        "enum" : [ "DRAFT", "COLLECTING", "READY", "PUBLISHED" ],
        "type" : "string"
      },
      "BatchStatusCounts" : {
        "description" : "Batch count per status; all is the sum",
        "properties" : {
          "COLLECTING" : {
            "format" : "int64",
            "type" : "integer"
          },
          "DRAFT" : {
            "format" : "int64",
            "type" : "integer"
          },
          "PUBLISHED" : {
            "format" : "int64",
            "type" : "integer"
          },
          "READY" : {
            "format" : "int64",
            "type" : "integer"
          },
          "all" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "all" ],
        "type" : "object"
      },
      "BatchUpdateRequest" : {
        "properties" : {
          "batchNo" : {
            "type" : [ "string", "null" ]
          },
          "producedFrom" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "producedTo" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "productionOrderNo" : {
            "type" : [ "string", "null" ]
          },
          "quantity" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          }
        },
        "type" : "object"
      },
...
      "ChainSummary" : {
        "properties" : {
          "approvedSteps" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalSteps" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "approvedSteps", "totalSteps" ],
        "type" : "object"
      },
...
      "FieldViolation" : {
        "description" : "Field errors: type urn:tekpas:problem:validation, or conflict on a unique field (code Unique)",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : [ "string", "null" ]
          },
          "params" : {
            "additionalProperties" : true,
            "example" : {
              "total" : 95
            },
            "type" : [ "object", "null" ]
          }
        },
        "required" : [ "code", "field" ]
      },
...
      "NextBatchNoResponse" : {
        "properties" : {
          "batchNo" : {
            "type" : "string"
          }
        },
        "required" : [ "batchNo" ],
        "type" : "object"
      },
      "PageResponseBatchResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/BatchResponse"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "content", "page", "size", "totalElements", "totalPages" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Digital Product Passport platform for textile manufacturers",
    "title" : "KozaPass API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/batches" : {
      "get" : {
        "operationId" : "listBatches",
        "parameters" : [ {
          "description" : "Search in batch no, production order no, product name and GTIN",
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "productId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/BatchStatus"
          }
        }, {
          "description" : "Batches whose supply chain has a step of this supplier (404 if the supplier is not in the company's network)",
          "in" : "query",
          "name" : "supplierId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Zero-based page",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "batchNo | productName | quantity | producedFrom | createdAt | updatedAt, optionally ,asc or ,desc (default createdAt,desc)",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseBatchResponse"
                }
              }
            },
            "description" : "One page of batches"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Invalid paging or sort parameter"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Supplier users have no access"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "supplierId is not in the company's supplier network"
          }
        },
        "summary" : "List batches with product and supply chain progress",
        "tags" : [ "batches" ]
      },
      "post" : {
        "description" : "Without batchNo the next KP-YYYY-MMDD-<letter> number of the day is assigned.",
        "operationId" : "createBatch",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BatchCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BatchResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Validation failed (e.g. Pattern on batchNo, DateRange)"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Read-only or supplier user"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "No such product in this company"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "The product already has a batch with this number"
          }
        },
        "summary" : "Create a DRAFT batch of a product",
        "tags" : [ "batches" ]
      }
    },
    "/api/v1/batches/next-batch-no" : {
      "get" : {
        "operationId" : "getNextBatchNo",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NextBatchNoResponse"
                }
              }
            },
            "description" : "Suggested number, e.g. KP-2026-1003-B"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Supplier users have no access"
          }
        },
        "summary" : "Suggest the next batch number of today (not reserved)",
        "tags" : [ "batches" ]
      }
    },
    "/api/v1/batches/status-counts" : {
      "get" : {
        "description" : "All batches of the company, independent of the list's search and filters",
        "operationId" : "getBatchStatusCounts",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BatchStatusCounts"
                }
              }
            },
            "description" : "Count per status, zero included"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Supplier users have no access"
          }
        },
        "summary" : "Count batches per status",
        "tags" : [ "batches" ]
      }
    },
    "/api/v1/batches/{id}" : {
      "delete" : {
        "operationId" : "deleteBatch",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Read-only or supplier user"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "No such batch in this company"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Not DRAFT (BATCH_NOT_DRAFT) or has a passport (BATCH_HAS_PASSPORT)"
          }
        },
        "summary" : "Delete a DRAFT batch without a passport",
        "tags" : [ "batches" ]
      },
      "get" : {
        "operationId" : "getBatch",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BatchResponse"
                }
              }
            },
            "description" : "The batch"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Supplier users have no access"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "No such batch in this company"
          }
        },
        "summary" : "A batch with its product and supply chain progress",
        "tags" : [ "batches" ]
      },
      "patch" : {
        "description" : "Missing fields stay unchanged; null clears productionOrderNo and the dates. Status and product cannot be changed here (such fields are ignored).",
        "operationId" : "updateBatch",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BatchUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BatchResponse"
                }
              }
            },
            "description" : "Updated"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Validation failed"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Read-only or supplier user"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "No such batch in this company"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Batch number taken, or changed after DRAFT"
          }
        },
        "summary" : "Change some fields of a batch",
        "tags" : [ "batches" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
}